requirements.txt
aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.2
frozenlist==1.4.0
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.2
magic-filter==1.0.12
Mako==1.2.4
MarkupSafe==2.1.3
multidict==6.0.4
psycopg2-binary==2.9.9
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
PyYAML==6.0.1
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
yarl==1.9.2

redis~=5.0.1